fix: merge handling-behavior tests to satisfy pkg-file-count limit
pkg/factory/validation exceeded the 15-file package cap after story 002;
fold work_type_handling_behavior tests into validation_test.go.
